fix общее

+++ main-service/src/main/java/ru/practicum/service/category/CategoryServiceImpl.java
@@ -6,10 +6,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.practicum.dto.category.CategoryDto;
 import ru.practicum.dto.category.NewCategoryDto;
+import ru.practicum.exception.ConflictException;
 import ru.practicum.mapper.CategoryMapper;
 import ru.practicum.model.Category;
 import ru.practicum.repository.CategoryRepository;
 import ru.practicum.exception.NotFoundException;
+import ru.practicum.repository.EventRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -19,6 +21,7 @@ import java.util.stream.Collectors;
 @Transactional(readOnly = true)
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
@@ -33,7 +36,9 @@ public class CategoryServiceImpl implements CategoryService {
         if (!categoryRepository.existsById(catId)) {
             throw new NotFoundException("Категория с ID=" + catId + " не найдена.");
         }
-        // ДОБАВИТЬ ПРОВЕРКУ ЧТО С КАТЕГОРИЕЙ НЕ СВЯЗАНО НИ ОДНОГО СОБЫТИЯ
+        if (eventRepository.existsByCategoryId(catId)) {
+            throw new ConflictException("Нельзя удалить категорию, с которой связаны события.");
+        }
         categoryRepository.deleteById(catId);
     }
 
@@ -48,7 +53,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public List<CategoryDto> getAllCategories(int from, int size) {
-        PageRequest page = PageRequest.of(from > 0 ? from / size : 0, size);
+        PageRequest page = PageRequest.of(from / size, size);
         return categoryRepository.findAll(page).stream()
                 .map(CategoryMapper::toCategoryDto)
                 .collect(Collectors.toList());
